pyproject.toml
[project]
name = "python-ai-teacher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "ruff>=0.13.0",
    "streamlit>=1.49.1",
]
